Re-implement glyph caching at rasterizer stage.
This provides a generic caching of shapes for rasterization - in a manner that
benefits all run-modes of the rasterizer (but currently cannot be used
threaded.).

This caches the coverage masks for shapes with less than 32x32 pixels, this is
done by hashing the flattened path in the rasterizer - before this is in-place
turned into edges - and scanlines are filled with or without vertical
super-sampling.

When compiled without optimization and instrumented the overhead of flattening
the path is much higher than when compiled with optimization (and thus
shallower call stacks). At the moment the earlier bailing for bezier
subdivision will cause clipped shapes to not be shareable between tiles in
deferred rendering.

This is a place-hodler that should work along with networked compression,
but when doing front-to-back randering, storing a hash along with bounds
in render-stream and caching at this level will avoid overhead computation.