ctx: add a file-system backed font backend
This is experimental and acts to direct the refactoring of the font
handling for experiments like dynamic subsetting. Right now; when
enabled - a hard-coded path of /tmp/ctx-font is used for online
no-caches used parsing from ctx protocol text files to used glyphs. Ctx
does not know that glyphs have changed - so re-renders should be forced
when this is used for font specimen proofing.

Each glyph is expected to be the glyphs related fragment in a ctx
drawlist starting with:

defineGlyph "a" 30.0

following each glyph, kerningPair commands where this glyph is the left
glyph follows. The tool ctx-font-split can be used to split a well
formed ctxf font drawlist to the individual files. Combined with
ctx-fontgen this permits converting TTF files to ctx font drawlists.