ctx: a85 disable z
The z becoming 4 0s makes the inplace expansion invalid, the
cases we encounter this. Not being able to do in-place expansion
is worse than incurring the overhead.

A longer term solution might be to have a buffer of up to a settable
minimum rescued bytes buffer, as data quickly will make the writes not
overlaps reads again.