add duktape as js engine
For now the duktype binding provides a mock web-like interface
and implements the HTML5 Canvas 2D context on top of ctx.

This provides the ability to run scripts intended to be run inside a
canvas in a webpage, perhaps full screen for a game to run as an
application/applet locally.