terminal: make ctx recording of tabs compiletime optional
This optimization makes the vector drawlist compositing/splicing more
faster when there is many, non-changing clients. For a single updating
client the impact of an additional drawlist copy is perceptible in
cpu/battery usage - maybe it should only be done for non
visible/floating windows or when there is more than one client?