ctx: optimize rasterization of strokes for single_lines segments
Keep track of first/last non-transparent pixel after generating coverage
for short lines. This is advantagous for diagonal lines when the
compositing format differs from the actual format, like with RGB565
formats. For these formats we might even want to opt in for this
optimization at a wider scope, it is kept as this potetial worst-case
scenarios as it adds undesired overhead/complexity for RGBA8 diretly.