ctx: implement optional fast low gradation slope AA mode
When the environment variable CTX_FAST_AA is set to "1" a
faster/in-development rasterizer is now used for the cases where the
simple rasterizers horizontal AA is insufficient. It is only really
valid when there are no edge crossings or edge aa overlap, these cases
occur seldom but with tuning it should only apply then.

The AA is currently limited to a fixed gradient for 2,3,4,5,6,7 etc
needed horizontal pixels of AA for the slope which is why we stretch the
AA of horizontal computation a bit further than 45 degrees.