ctx: make it possible to build event handling without unix bits
This is done by adding a CTX_TERMINAL_EVENTS define, which controls
conditional compilation of these bits, taking over the role of
CTX_EVENTS, and iheriting its value when not configured.

This is for embedded targets earlier disabling events was done to
disable these bits, to be renamed when better name comes up.